Include nextcloud service in trusted domains for metrics exporter (#483)
* Include service in trusted domains for metrics

When nextcloud-exporter is used by enabling metrics, it fails to fetch
the serverinfo with 400 response code. This is related to the switch
from ingress to local service endpoint for nextcloud-exporter.
To fix this, in addition to .Values.nextcloud.host, the nextcloud
service also need to be included in the trusted domains.

Signed-off-by: Sunny <darkowlzz@protonmail.com>

* Update charts/nextcloud/templates/_helpers.tpl - use full service name and fix spacing for trusted domains

Co-authored-by: WrenIX <133280015+wrenix@users.noreply.github.com>
Signed-off-by: JesseBot <jessebot@linux.com>

---------

Signed-off-by: Sunny <darkowlzz@protonmail.com>
Signed-off-by: JesseBot <jessebot@linux.com>
Co-authored-by: JesseBot <jessebot@linux.com>
Co-authored-by: WrenIX <133280015+wrenix@users.noreply.github.com>
Signed-off-by: switzerchees <patrickmichel96@yahoo.de>